chore: use tsup
This replaces `tshy` with [tsup](https://tsup.egoist.dev) because I wanted to try it.  Unfortunately, it entrains `@microsoft/api-extractor` to use `tsup`'s own declaration-generating facilities because it is a [pain in the ass](https://mayank.co/blog/dual-packages/) to make `tsc` do this.  On the plus side, it seems to take care of broken global Node.js incompatibilities for us (so far).

It also migrates `@stylistic/eslint-plugin-js` & `@stylistic/eslint-plugin-ts` to `@stylistic/eslint-plugin`.